Create Fsm serial + datapath
Update to Fsm serial by adding a register to save data being send. 

Implement by shifting the input signal to the corresponding location in the register. The output use input byte from the end of the signal to the beginning. As a result, the first input is in 0 index of [7:0] and 2nd input is index ...