Create Fsm serial + datapath + odd parity
Create a module that output whether an odd number of 1s is received.

The input is now 9 bits long with a parity bit at the end, the output is still 9 bit.

add a parse checking state to prevent the saving_bit register from receiving more than 8 bit. the parse check can also work as a state determining whether error occur. 

output done is 0 if the input is even