Create Moore Fsm(q5a)
least_sig: check if the LSB is 0 because inverse = 1 and plus 1 will leave remainder.

one_remainder: remain as long as input = 0 to add remainder 1. exit when input = 1.

inversing: not all input. note that previous state is one_remainder to add the final remainder 

resetting to least_sig because there is no output logic. The program won't output any value