Create reviews2015 fsmonehot
The input state will be tested with non-onehot so we need a for loop to check all the state will be use

output done, counting and shift_ena depend on current state from input

all other output depend on next_state

For ease of checking, I uses shift to place 1 at the according next_state. It might be possible to use s? (next_state[1] = 1 : next_state[2] = 1) to do the same thing. IDK if it is legal