Create gshare
Using the counter 2bc and history shift for the base of the code.

Looking at the shift history code, the train_mispredicted have priority over predict_valid. On the cycle,  we will shift in train_taken when train_mispredicted is true. When train_mispredicted is false and predict_valid, shift in predict_taken to predict_history.

For the second part, I use counter 2bc. The question mention pht uses hashed of PC and history, but is vague on which, of train.
If train taken, we increase the pht at index by 1 else decrease by 0.
Note: pht on reset = 2'b01 for all the array and max at 3, min at 0